backend/requirements.txt
fastapi
uvicorn[standard]
requests
azure-identity
azure-storage-blob
pandas
pyarrow
pydantic-settings
